changes to scripts
added code to grab stats from the class stats and use it

+++ 3D Group Project/Assets/Scripts/Projectile.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Projectile : MonoBehaviour
 {
-    private int damage = 1;
+    private float damage = 1;
 
     private void Start()
     {
